Moved Signal client tests
