+++ signal/client/client_suite_test.go
@@ -1,4 +1,4 @@
-package signal_test
+package client
 
 import (
 	. "github.com/onsi/ginkgo"
